test: Remove and replace JiraServerPerformanceTestV2 with original test
